Added default handler with no business logic. EGSP handlers to be introduced in a separate PR.
